`main` rebase fixes
