+++ Sources/WasmKit/Execution/ComponentModel/Lift+Lower.swift
@@ -1410,7 +1410,7 @@
             default:
                 // For complex types, we'll need heap-based ABI
                 // For now, return empty to avoid crashes - these will fail at runtime
-                #warning("Complex types not supported in `flattenComponentValueType`")
+                // TODO: Complex types not supported in `flattenComponentValueType`
                 return []
             }
         }
