+++ Tests/WATTests/EncoderTests.swift
@@ -104,7 +104,7 @@ struct EncoderTests {
         }
 
         switch module.source {
-        case .text(var wat):
+        case .text(let wat):
             #expect(throws: (any Error).self, diagnostic()) {
                 _ = try wat.encode()
                 recordFail()
@@ -433,7 +433,7 @@ struct EncoderTests {
                 nameBytes = section.bytes
             }
         }
-        let sectionBytes = try #require(Array(nameBytes ?? []))
+        let sectionBytes = Array(nameBytes ?? [])
         let nameParser = NameSectionParser(
             stream: StaticByteStream(bytes: sectionBytes)
         )
